+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Changelog for Project Utilities
 * Adding additional documentation via `README.md` files so that users can view them via GitHub sources.
 * Updated the composer dependencies.
 * Added `ObsidianMoon\ProjectUtilities\Traits\WordPress\Settings\SettingsRepositoryTrait`.
+* Added `ObsidianMoon\ProjectUtilities\Interfaces\WordPress\Taxonomy\TaxonomyValidationInterface`.
 
 ### Version 0.5.2